+++ examples/example_data/relion_3d_simple.star
@@ -1,0 +1,65 @@
+# Created by the starfile python package (version 0.3.2) on 12/11/2020 16:45:12
+
+data_
+
+loop_
+_rlnCoordinateX #1
+_rlnCoordinateY #2
+_rlnCoordinateZ #3
+_rlnAngleRot #4
+_rlnAngleTilt #5
+_rlnAnglePsi #6
+_rlnMicrographName #7
+0.29950025	2.98501250	0.10000000	84.27042205	90.00000000	-84.27042205	test_tomo
+1.39778850	2.65446554	0.48468481	62.22960572	90.00000000	-62.22960572	test_tomo
+2.29176692	1.93592469	0.86936963	40.18878940	90.00000000	-40.18878940	test_tomo
+2.85076582	0.93441653	1.25405444	18.14797307	90.00000000	-18.14797307	test_tomo
+2.99307830	-0.20367201	1.63873926	-3.89284326	90.00000000	3.89284326	test_tomo
+2.69790305	-1.31199053	2.02342407	-25.93365958	90.00000000	25.93365958	test_tomo
+2.00838479	-2.22854000	2.40810889	-47.97447591	90.00000000	47.97447591	test_tomo
+1.02530798	-2.81935162	2.79279370	-70.01529224	90.00000000	70.01529224	test_tomo
+-0.10763449	-2.99806851	3.17747852	-92.05610856	90.00000000	92.05610856	test_tomo
+-1.22484440	-2.73856827	3.56216333	-114.09692489	90.00000000	114.09692489	test_tomo
+-2.16302311	-2.07878113	3.94684815	-136.13774122	90.00000000	136.13774122	test_tomo
+-2.78504034	-1.11514586	4.33153296	-158.17855754	90.00000000	158.17855754	test_tomo
+-2.99997801	0.01148636	4.71621778	179.78062613	90.00000000	-179.78062613	test_tomo
+-2.77641944	1.13643966	5.10090259	157.73980980	90.00000000	-157.73980980	test_tomo
+-2.14704139	2.09528357	5.48558741	135.69899348	90.00000000	-135.69899348	test_tomo
+-1.20383785	2.74786725	5.87027222	113.65817715	90.00000000	-113.65817715	test_tomo
+-0.08467357	2.99880483	6.25495704	91.61736082	90.00000000	-91.61736082	test_tomo
+1.04686716	2.81141764	6.63964185	69.57654450	90.00000000	-69.57654450	test_tomo
+2.02539099	2.21309542	7.02432667	47.53572817	90.00000000	-47.53572817	test_tomo
+2.70787054	1.29129282	7.40901148	25.49491184	90.00000000	-25.49491184	test_tomo
+2.99455017	0.18074649	7.79369629	3.45409552	90.00000000	-3.45409552	test_tomo
+2.84352693	-0.95621892	8.17838111	-18.58672081	90.00000000	18.58672081	test_tomo
+2.27687535	-1.95341717	8.56306592	-40.62753713	90.00000000	40.62753713	test_tomo
+1.37742090	-2.66509130	8.94775074	-62.66835346	90.00000000	62.66835346	test_tomo
+0.27663368	-2.98721841	9.33243555	-84.70916979	90.00000000	84.70916979	test_tomo
+-0.86458811	-2.87271429	9.71712037	-106.74998611	90.00000000	106.74998611	test_tomo
+-1.87943609	-2.33831563	10.10180518	-128.79080244	90.00000000	128.79080244	test_tomo
+-2.61957351	-1.46213358	10.48649000	-150.83161877	90.00000000	150.83161877	test_tomo
+-2.97681708	-0.37223662	10.87117481	-172.87243509	90.00000000	172.87243509	test_tomo
+-2.89894975	0.77206887	11.25585963	165.08674858	90.00000000	-165.08674858	test_tomo
+-2.39735313	1.80352377	11.64054444	143.04593225	90.00000000	-143.04593225	test_tomo
+-1.54534383	2.57136393	12.02522926	121.00511593	90.00000000	-121.00511593	test_tomo
+-0.46745705	2.96335686	12.40991407	98.96429960	90.00000000	-98.96429960	test_tomo
+0.67875628	2.92220634	12.79459889	76.92348327	90.00000000	-76.92348327	test_tomo
+1.72575819	2.45392719	13.17928370	54.88266695	90.00000000	-54.88266695	test_tomo
+2.52051210	1.62696613	13.56396852	32.84185062	90.00000000	-32.84185062	test_tomo
+2.94685160	0.56219714	13.94865333	10.80103429	90.00000000	-10.80103429	test_tomo
+2.94246017	-0.58474623	14.33333814	-11.23978203	90.00000000	11.23978203	test_tomo
+2.50797967	-1.64621929	14.71802296	-33.28059836	90.00000000	33.28059836	test_tomo
+1.70691661	-2.46707027	15.10270777	-55.32141469	90.00000000	55.32141469	test_tomo
+0.65635953	-2.92731825	15.48739259	-77.36223101	90.00000000	77.36223101	test_tomo
+-0.49013530	-2.95969042	15.87207740	-99.40304734	90.00000000	99.40304734	test_tomo
+-1.56498878	-2.55945504	16.25676222	-121.44386367	90.00000000	121.44386367	test_tomo
+-2.41109336	-1.78511311	16.64144703	-143.48467999	90.00000000	143.48467999	test_tomo
+-2.90477689	-0.74984748	17.02613185	-165.52549632	90.00000000	165.52549632	test_tomo
+-2.97387939	0.39502073	17.41081666	172.43368735	90.00000000	-172.43368735	test_tomo
+-2.60830039	1.48215015	17.79550148	150.39287103	90.00000000	-150.39287103	test_tomo
+-1.86147530	2.35263888	18.18018629	128.35205470	90.00000000	-128.35205470	test_tomo
+-0.84256490	2.87925067	18.56487111	106.31123838	90.00000000	-106.31123838	test_tomo
+0.29950025	2.98501250	18.94955592	84.27042205	90.00000000	-84.27042205	test_tomo
+
+
+
